rescuebox/pyproject.toml
[tool.poetry]
name = "rescuebox"
version = "0.1.0"
description = ""
authors = ["Jagath Jai Kumar <jagath.jaikumar@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
typer = "^0.12.5"
requests = "^2.32.3"
pyyaml = "^6.0.2"
ollama = "^0.3.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
